シャドウマップを消し消し

+++ Game_Dxlib/Src/Ball.h
@@ -17,8 +17,8 @@ public:
 private:
 	int color;
 
-	float rote;
-	float speed;
+	int rote;
+	int speed;
 
 	GSvector3 Power;
 };
